zstandard-compare: Turn off UBSAN for now
This seems to have been the cause of the unknown instruction (the instruction was UB1). In theory this means we were finding undefined behavior within the zstd C library, but I haven't been able to reproduce it after compiling zstd normally. Will need to look into it more, this is a workaround to get the fuzzer running.