More zstd UBSAN avoidance
This avoids a bug when the dest ptr is `intMax(usize)` and the dest len is 0. Writing the equivalent program in C would not trigger this particular bug because `malloc` always gives a real pointer even if the length requested is 0. Zig's Allocator does not do a real allocation if the requested size is zero, though, so we were passing in `intMax(usize)` as the dest ptr which was triggered undefined behavior in zstd.

Even with this mitigation, we still need to keep UBSAN turned off because we are hitting more real undefined behavior in ZSTD_prefetchMatch for some inputs (but it doesn't seem to trigger buggy behavior so we can ignore it for now).