fuzz-tar: Put tmp dirs into a subdir for easier cleanup
Panics bypass the deleteTree, so there will be directories left over when crashes are found. However, there seems to be way more directories left over than there are crashes found, so there seems to be something else going on but I can't seem to reproduce that to know why. It doesn't seem to be the case that deleteTree is being run but failing (tried `catch panic` in the deleteTree call but didn't hit it while fuzzing AFAICT), so it's either `deleteTree` silently failing but returning success, or something else very strange is going on.